Move Selenium build and quit to helper function
