+++ __test__/e2e/util/setup.js
@@ -1,1 +1,2 @@
+// This script will execute before the entire end to end run
 jest.setTimeout(60000)
